Move KeyValueSpout related things to storm.builder
